Move SlotsManipulator into controller
This separation dated from when instant and deferred booking were
separate. It's obviously not needed now.
