+++ spec/controllers/slots_controller_spec.rb
@@ -21,11 +21,6 @@ RSpec.describe SlotsController, type: :controller do
   it_behaves_like "a browser without a session present"
   it_behaves_like "a session timed out"
 
-  it "sets up the flow" do
-    expect(controller.this_path).to eq(edit_slots_path)
-    expect(controller.next_path).to eq(edit_visit_path)
-  end
-
   it "permits up to 3 slots to be selected" do
     expect(controller.max_slots).to eq(3)
   end
@@ -35,17 +30,17 @@ RSpec.describe SlotsController, type: :controller do
       {
         visit: {
           slots: [
-                  {
-                    slot: '2013-01-01-1345-2000'
-                  }
-                 ]
+            {
+              slot: '2013-01-01-1345-2000'
+            }
+          ]
         }
       }
     end
 
     it "permits us to select a time slot" do
       post :update, slots_hash
-      expect(response).to redirect_to(controller.next_path)
+      expect(response).to redirect_to(edit_visit_path)
     end
   end
 
@@ -58,7 +53,7 @@ RSpec.describe SlotsController, type: :controller do
 
     it "prompts us to retry" do
       post :update, slots_hash
-      expect(response).to redirect_to(controller.this_path)
+      expect(response).to redirect_to(edit_slots_path)
     end
   end
 
@@ -71,7 +66,7 @@ RSpec.describe SlotsController, type: :controller do
 
     it "accepts the submission" do
       post :update, slots_hash
-      expect(response).to redirect_to(controller.next_path)
+      expect(response).to redirect_to(edit_visit_path)
     end
   end
 
@@ -84,7 +79,7 @@ RSpec.describe SlotsController, type: :controller do
 
     it "accepts the submission" do
       post :update, slots_hash
-      expect(response).to redirect_to(controller.next_path)
+      expect(response).to redirect_to(edit_visit_path)
     end
   end
 
@@ -97,7 +92,7 @@ RSpec.describe SlotsController, type: :controller do
 
     it "prompts us to retry" do
       post :update, slots_hash
-      expect(response).to redirect_to(controller.this_path)
+      expect(response).to redirect_to(edit_slots_path)
       expect(session[:visit].errors[:slots]).not_to be_nil
     end
   end
